Draw sphere wires

Draw sphere wires.

Usage

draw_sphere_wires(center_pos, radius, rings, slices, color)

Arguments

center_pos

A numeric vector of length 3 or a numeric matrix of width 3.

radius

A number or a vector of numbers.

rings

An integer or a vector of integers.

slices

An integer or a vector of integers.

color

A color or a list of colors.

Note

This function is vectorized to allow for faster drawing. This function has been auto-generated from the following Raylib function definition:

void DrawSphereWires(Vector3 centerPos, float radius, int rings, int slices, Color color);

See also