Load Sound Alias

Create a new sound that shares the same sample data as the source sound, does not own the sound data.

Usage

load_sound_alias(source)

Arguments

source

A sound.

Value

A sound

Note

This function has been auto-generated from the following Raylib function definition:

Sound LoadSoundAlias(Sound source);

See also

Examples

if (FALSE) {
load_sound_alias(load_sound("sound.wav"))
}